backend/dev-requirements.txt
# Development Python requirements

-r requirements.txt

django-dynamic-fixture
flake8
ipython
pip-tools
pyinotify
pylint
pytest
pytest-catchlog
pytest-django
redislite
